# swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "A list of the currently provided Interfaces to interact with FairOS decentralised file system(dfs), implementing user, pod, file system, key value store and document store",
        "title": "FairOS-dfs server",
        "contact": {
            "name": "Sabyasachi Patra",
            "email": "sabyasachi@datafund.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "v0.0.0"
    },
    "paths": {
...
        "/v1/file/receive": {
            "get": {
                "description": "FileReceiveHandler is the api handler to receive a file in a given pod",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "file"
                ],
                "summary": "Receive a file",
                "parameters": [
                    {
                        "type": "string",
                        "description": "pod name",
                        "name": "podName",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "sharing reference",
                        "name": "sharingRef",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "file location",
                        "name": "dirPath",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "cookie parameter",
                        "name": "Cookie",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.FileSharingReference"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.response"
                        }
                    }
                }
            }
        },
        "/v1/file/receiveinfo": {
            "get": {
                "description": "FileReceiveInfoHandler is the api handler to receive a file info",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "file"
                ],
                "summary": "Receive a file info",
                "parameters": [
                    {
                        "type": "string",
                        "description": "sharing reference",
                        "name": "sharingRef",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "cookie parameter",
                        "name": "Cookie",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/user.ReceiveFileInfo"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.response"
                        }
                    }
                }
            }
        },
...
        "/v1/file/share": {
            "post": {
                "description": "FileShareHandler is the api handler to share a file from a given pod",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "file"
                ],
                "summary": "Share a file",
                "parameters": [
                    {
                        "description": "file share request params",
                        "name": "file_share_request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.FileShareRequest"
                        }
                    },
                    {
                        "type": "string",
                        "description": "cookie parameter",
                        "name": "Cookie",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.FileSharingReference"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.response"
                        }
                    }
                }
            }
        },
        "/v1/file/stat": {
            "get": {
                "description": "FileStatHandler is the api handler to get the information of a file",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "file"
                ],
                "summary": "Info of a file",
                "parameters": [
                    {
                        "type": "string",
                        "description": "pod name",
                        "name": "podName",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "file path",
                        "name": "filePath",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "cookie parameter",
                        "name": "Cookie",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/file.Stats"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "api.FileShareRequest": {
            "type": "object",
            "properties": {
                "destUser": {
                    "type": "string"
                },
                "filePath": {
                    "type": "string"
                },
                "podName": {
                    "type": "string"
                }
            }
        },
        "api.FileSharingReference": {
            "type": "object",
            "properties": {
                "fileSharingReference": {
                    "type": "string"
                }
            }
        },
...
        "api.response": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
...
        "file.Blocks": {
            "type": "object",
            "properties": {
                "compressedSize": {
                    "type": "string"
                },
                "reference": {
                    "type": "string"
                },
                "size": {
                    "type": "string"
                }
            }
        },
...
        "file.Stats": {
            "type": "object",
            "properties": {
                "accessTime": {
                    "type": "string"
                },
                "blockSize": {
                    "type": "string"
                },
                "blocks": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/file.Blocks"
                    }
                },
                "compression": {
                    "type": "string"
                },
                "contentType": {
                    "type": "string"
                },
                "creationTime": {
                    "type": "string"
                },
                "fileName": {
                    "type": "string"
                },
                "filePath": {
                    "type": "string"
                },
                "fileSize": {
                    "type": "string"
                },
                "modificationTime": {
                    "type": "string"
                },
                "podName": {
                    "type": "string"
                }
            }
        },
...
        "user.ReceiveFileInfo": {
            "type": "object",
            "properties": {
                "blockSize": {
                    "type": "string"
                },
                "compression": {
                    "type": "string"
                },
                "contentType": {
                    "type": "string"
                },
                "destAddress": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "numberOfBlocks": {
                    "type": "string"
                },
                "sharedTime": {
                    "type": "string"
                },
                "size": {
                    "type": "string"
                },
                "sourceAddress": {
                    "type": "string"
                }
            }
        },
...
    }
}